Add license headers to lit test python files
This patch adds license headers to the lit test python files for the
convert_bhive_to_llvm_exegesis_input script that should have been added
on in the initial patch. Additionally, this patch changes the import in
one file to just lit.main instead of using a from to comply with the
google python style guide.