Move loop register into BlockAnnotations (#166)
convert_bhive_to_llvm_exegesis_input before this patch stored the
loop_register separately. Having the annotators add the loop register to
the annotations gives more flexibility down the line for exactly
specifying how the block was executed in the annotator/should be
benchmarked. Additionally, a register chosen for use as a loop counter
is fundamentally an annotation, so placing it there makes more sense
from the abstraction perspective.